Moved color to its own file
